packages/cosmos-oss/pyproject.toml
# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
# SPDX-License-Identifier: Apache-2.0
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
# http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cosmos-oss"
dynamic = ["version"]
requires-python = ">=3.10"
dependencies = [
  "albumentations>=2.0.8",
  "attrs>=25.4.0",
  "av>=16.0.1",
  "better-profanity>=0.7.0",
  "boto3>=1.40.58",
  "click>=8.3.0",
  "decord>=0.6.0; platform_machine != 'aarch64'",
  "diffusers>=0.35.2",
  "einops>=0.8.1",
  "ftfy>=6.3.1",
  "fvcore>=0.1.5.post20221221",
  "fastparquet>=2024.11.0",
  "h11>=0.16.0",
  "huggingface-hub>=0.36.0",
  "hydra-core>=1.3.2",
  "imageio[pyav,ffmpeg]>=2.37.0",
  "iopath>=0.1.10",
  "ipdb>=0.13.13",
  "loguru>=0.7.3",
  "mediapy>=1.2.4",
  "megatron-core>=0.14.0",
  "ml-dtypes>=0.5.3",
  "modelscope>=1.31.0",
  "multi-storage-client[boto3,google-cloud-storage,fsspec,observability-otel,msal]>=0.32.0",
  "nltk>=3.9.2",
  "numpy",
  "numpydantic>=1.7.0",
  "nvidia-ml-py>=13.580.82",
  "omegaconf>=2.3.0",
  "opencv-python>=4.11.0.86",
  "pandas>=2.3.3",
  "pyarrow>=19.0.0",
  "peft>=0.17.1",
  "pillow>=11.3.0",
  "protobuf>=6.33.0",
  "pycocotools>=2.0.10",
  "pydantic>=2.11.7",
  "pyinstrument>=5.1.1",
  "pytest>=8.4.2",
  "pyyaml>=6.0.3",
  "qwen-vl-utils[decord]>=0.0.14",
  "retinaface-py>=0.0.2",
  "safetensors>=0.6.2",
  "scikit-image>=0.25.2",
  "scipy>=1.15.3",
  "sentencepiece>=0.2.1",
  "setuptools>=80.9.0",
  "termcolor>=3.1.0",
  "tqdm>=4.67.1",
  "transformers==4.51.3",
  "typing-extensions>=4.14.1",
  "tyro>=0.9.35",
  "wandb>=0.22.2",
  "webdataset>=1.0.2",
]

[project.optional-dependencies]
dev = [
  "hatch>=1.15.1",
  "pyrefly==0.39.2",
  "pytest>=8.4.1",
  "pytest-custom_exit_code>=0.3.0",
  "pytest-cov>=6.2.1",
  "pytest-env>=1.1.5",
  "pytest-instafail>=0.5.0",
  "pytest-regressions>=2.8.3",
  "pytest-xdist>=3.8.0",
  "ruff==0.12.7",
]
gradio = ["cosmos-gradio==0.1.0"]
# Match nvcr.io/nvidia/pytorch:25.10-py3
cu128_torch27 = [
  "cosmos-cuda", # Sentinel package
  "decord>=0.6.0",
  "flash-attn==2.7.3",
  "natten==0.21.0",
  "transformer-engine==2.2.0",
  # Torch
  "torch==2.7.0",
  # Dependencies constrained by torch
  "torchvision",
  "triton",
  "xformers; platform_machine != 'aarch64'",
  # Dependencies determined from 'uv pip install "torch==2.7.0" --index-url https://download.pytorch.org/whl/cu128'
  # Issue: https://github.com/astral-sh/uv/issues/14237
  "nvidia-cublas-cu12==12.8.3.14",
  "nvidia-cuda-cupti-cu12==12.8.57",
  "nvidia-cuda-nvrtc-cu12==12.8.61",
  "nvidia-cuda-runtime-cu12==12.8.57",
  "nvidia-cudnn-cu12==9.7.1.26",
  "nvidia-cufft-cu12==11.3.3.41",
  "nvidia-cufile-cu12==1.13.0.11",
  "nvidia-curand-cu12==10.3.9.55",
  "nvidia-cusolver-cu12==11.7.2.55",
  "nvidia-cusparse-cu12==12.5.7.53",
  "nvidia-cusparselt-cu12==0.6.3",
  "nvidia-nccl-cu12==2.26.2",
  "nvidia-nvjitlink-cu12==12.8.61",
  "nvidia-nvtx-cu12==12.8.55",
]
# Match nvcr.io/nvidia/pytorch:25.10-py3
cu130_torch29 = [
  "cosmos-cuda", # Sentinel package
  "flash-attn==2.7.4.post1",
  "natten==0.21.0",
  "transformer-engine==2.8.0",
  # Torch
  "torch==2.9.0",
  # Dependencies constrained by torch
  "torchvision",
  "triton",
  "xformers",
  # Dependencies determined from 'uv pip install "torch==2.9.0" --index-url https://download.pytorch.org/whl/cu130'
  # Issue: https://github.com/astral-sh/uv/issues/14237
  "nvidia-cublas==13.0.0.19",
  "nvidia-cuda-cupti==13.0.48",
  "nvidia-cuda-nvrtc==13.0.48",
  "nvidia-cuda-runtime==13.0.48",
  "nvidia-cudnn-cu13==9.13.0.50",
  "nvidia-cufft==12.0.0.15",
  "nvidia-cufile==1.15.0.42",
  "nvidia-curand==10.4.0.35",
  "nvidia-cusolver==12.0.3.29",
  "nvidia-cusparse==12.6.2.49",
  "nvidia-cusparselt-cu13==0.8.0",
  "nvidia-nccl-cu13==2.27.7",
  "nvidia-nvjitlink==13.0.39",
  "nvidia-nvshmem-cu13==3.3.24",
  "nvidia-nvtx==13.0.39",
]

[tool.uv]
conflicts = [
  [
    {extra = "cu128_torch27"},
    {extra = "cu130_torch29"},
  ],
]
override-dependencies = ["numpy>=2.0.0"]

[tool.uv.sources]
cosmos-cuda = [{ workspace = true }]
cosmos-gradio = [{ workspace = true }]
# https://docs.astral.sh/uv/guides/integration/pytorch/#configuring-accelerators-with-optional-dependencies
decord = [
  { index = "cosmos-cu128-torch27", extra = "cu128_torch27" },
]
flash-attn = [
  { index = "cosmos-cu128-torch27", extra = "cu128_torch27" },
  { index = "cosmos-cu130-torch29", extra = "cu130_torch29" },
]
natten = [
  { index = "cosmos-cu128-torch27", extra = "cu128_torch27" },
  { index = "cosmos-cu130-torch29", extra = "cu130_torch29" },
]
transformer-engine = [
  { index = "cosmos-cu128-torch27", extra = "cu128_torch27" },
  { index = "cosmos-cu130-torch29", extra = "cu130_torch29" },
]
torch = [
  { index = "cosmos-cu128-torch27", extra = "cu128_torch27" },
  { index = "cosmos-cu130-torch29", extra = "cu130_torch29" },
]
torchvision = [
  { index = "cosmos-cu128-torch27", extra = "cu128_torch27" },
  { index = "cosmos-cu130-torch29", extra = "cu130_torch29" },
]
triton = [
  { index = "cosmos-cu128-torch27", extra = "cu128_torch27" },
  { index = "cosmos-cu130-torch29", extra = "cu130_torch29" },
]
xformers = [
  { index = "cosmos-cu128-torch27", extra = "cu128_torch27" },
  { index = "cosmos-cu130-torch29", extra = "cu130_torch29" },
]

[[tool.uv.index]]
name = "cosmos-cu128-torch27"
url = "https://nvidia-cosmos.github.io/cosmos-dependencies/v1.2.0/cu128_torch27/simple"
explicit = true

[[tool.uv.index]]
name = "cosmos-cu130-torch29"
url = "https://nvidia-cosmos.github.io/cosmos-dependencies/v1.2.0/cu130_torch29/simple"
explicit = true

[tool.hatch.version]
path = "cosmos_oss/__about__.py"

[tool.hatch.build.targets.sdist]
packages = [
  "cosmos_oss",
]

[tool.hatch.build.targets.wheel]
packages = [
  "cosmos_oss",
]
exclude = [
  "*_test.py",
]
